% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_QTcf}
\alias{sim_QTcf}
\title{Simulated QTcf exposure-response data}
\description{
This data frame contains simulated QTcf exposure-response data for 15 patients sampled at 8 different times after receiving a hypothetical drug.
}

